Simplify the build
This change gets rid of the need for the on_programs_changed shell
script. We can just have the user directly edit programs.make if they
want to build new programs. If the user hasn't created such a file, then
we default to just building the compiler.

If the user wants to build the compiler and a program called hello, then
they can make a programs.make file with the following contents and the
build system will figure out the rest.

programs = 84 hello